Fix duplicate UniqueIdFormat warnings
